Crate conjure_cp_rule_macros

Macros

register_rule_set

Register a rule set with the given name, dependencies, and metadata.Example

Attribute Macros

register_rule

Register a rule with the given rule sets and priorities.